feat(data_alignment) clean

+++ terraform/scripts/destroy_infra.ps1
@@ -1,4 +1,21 @@
 
 # This script will destroy the infrastructure created by the init_infra.ps1 script
-terraform plan -destroy -out main.destroy.tfplan
-terraform apply main.destroy.tfplan
+
+
+param([String]$env = "dev")
+
+
+if ("dev" -eq $env)
+{
+    echo "Start destroying infrastructure of the DEV environment..."
+    terraform plan -destroy -var-file dev.tfvars -out main.destroy.tfplan
+    terraform apply main.destroy.tfplan
+}
+
+if ("prd" -eq $env)
+{
+    echo "Start destroying infrastructure of the DEV environment..."
+    terraform plan -destroy -var-file prd.tfvars -out main.destroy.tfplan
+    terraform apply main.destroy.tfplan
+}
+
